FW/Win32/child_debug: detect the boundaries of the application
Given ASLR, we don't know that the pointers in the child process will
match those of the parent (they usually won't). With VirtualQuery() we
could obtain for any region, but a) I don't know how to obtain the name
of the image loaded at that location and b) we don't want to do a lot of
work inside the crash handler, after we've crashed.

So let's hope that the majority of crashes that are interesting to
obtain reproductions from are from the executable itself.

It now prints something like:
```yaml
    - { level: error, text: 'E> Received exception 0xc0000005 (Access violation), RIP = 0x7ff74a9e98dc (base+0x298dc), CR2 = 0x3d0 access=R' }
```

And the symbol table contains:
```
00000001400298d0 t force_memory_load(unsigned long long)
```

That address printed by nm(1) is has the image base applied; from the header:
```
ImageBase               0000000140000000
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>