FW/InterruptMonitor: fix parsing /proc/interrupts with offlined CPUs
When some CPUs are offlined by the system adminstrator via
`/sys/devices/system/cpu/cpuN/online`, the column numbers and CPU
numbers in `/proc/interrupts` will no longer match.

For example, if I turn off CPUs 10 to 19 on a 14-core system, the header
will be:
```
            CPU0       CPU1       CPU2       CPU3       CPU4       CPU5       CPU6       CPU7       CPU8       CPU9       CPU20      CPU21      CPU22      CPU23      CPU24      CPU25      CPU26      CPU27
```

So we must parse the first line to create tha mapping and we must do so
every time we read this file, in case it changed. If it did, the
mce_check will likely report an error.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>