CI/Bats: redirect output of the yq run to /dev/null
I don't know how this passed before, but I forgot the `sandstone_yq`
shell function passes the tool's output to its stdout, so Bats would log
it. That is fine when running in the `@test` but not if we are trying to
capture another shell function's output.

We were getting a very cryptic
```
opendcdiag/bats/sanity-check/21-slicing.bats: line 123: set: --: invalid option
```

Because the output of the `cpuset_two_modules` function was:
```
--- THIS IS AN UNOPTIMIZED BUILD: DON'T TRUST TEST TIMING! command-line: 'opendcdiag --on-crash=core --on-hang=kill -Y -o - --cpuset=t0 --disable=*' ...
```

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>